+++ functions.py
@@ -5,3 +5,22 @@ def add(a, b):
 def minus(a, b):
     """Subtract b from a."""
     return a - b
+
+def divide(a, b):
+  """a divide b"""
+    return a / b
+  
+def share(a, b):
+  """a share b """"
+    return a // b
+  
+def modulo(a, b):
+    """a mod b"""
+    if b == 0:
+        print("Can't operate!!")
+    else:
+        return a % b
+      
+def multiply(a, b):
+    """Multiply two numbers"""
+    return a * b