+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+----------------------
+
 ## [1.0.0] - 2025-05-11
 
 ### Added
@@ -11,3 +13,10 @@
 - Set up GitHub Actions workflows for automated testing and code review.
 - Created `requirements.txt` for managing project dependencies.
 - Provided `README.md` with basic project information and setup instructions.
+## [Unreleased]
+### Added
+
+- Added divide function in `functions.py`
+- Added share function in `functions.py`
+- Added modulo function in `functions.py`
+- Added multiply function in `functions.py`